Removes column sorting. Adds addTransaction scrolling.
Column sorting was causing issues on sorting amount. The amounts are always sorted as strings and not numbers. Also, more research is needed on resetting sorting to the default sort which is on a hidden field.

When Adding a trainsaction, the page now scrolls smoothly to the newly added row and focuses on the vendor field.