Changes the theme to dark (Superhero)
Chagnes the bootstrap meteor package used to nemo64:bootstrap. This is an uncomplied less version of bootstrap and is customizible. Changed the theme of the site to superhero and makes a few fixes to the layout and other styles to make it look right. Sets the undo timer to 6 seconds.